added many-to-one relation to certifaces and courses

+++ src/main/java/gr/hua/dit/ds/reference/letter/service/controller/StudentController.java
@@ -44,13 +44,13 @@ public class StudentController {
     }
 
     @GetMapping("/students/edit/{id}")
-    public String editStudentForm(@PathVariable Long id, Model model) {
+    public String editStudentForm(@PathVariable int id, Model model) {
         model.addAttribute("student", studentService.getStudentById(id));
         return "edit_student";
     }
 
     @PostMapping("/students/{id}")
-    public String updateStudent(@PathVariable Long id,
+    public String updateStudent(@PathVariable int id,
                                 @ModelAttribute("student") Student student,
                                 Model model) {
 
@@ -72,7 +72,7 @@ public class StudentController {
     // handler method to handle delete student request
 
     @GetMapping("/students/{id}")
-    public String deleteStudent(@PathVariable Long id) {
+    public String deleteStudent(@PathVariable int id) {
         studentService.deleteStudentById(id);
         return "redirect:/students";
     }
