+++ src/main/java/gr/hua/dit/ds/reference/letter/service/service/impl/TeacherServiceImpl.java
@@ -31,7 +31,7 @@ public class TeacherServiceImpl implements TeacherService {
         }
 
     @Override
-    public Teacher getTeacherById(Long id) {
+    public Teacher getTeacherById(int id) {
             return teacherRepository.findById(id).get();
         }
 
@@ -41,7 +41,7 @@ public class TeacherServiceImpl implements TeacherService {
         }
 
     @Override
-    public void deleteTeacherById(Long id) {
+    public void deleteTeacherById(int id) {
         teacherRepository.deleteById(id);
         }
 }
